// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/disposition.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type Disposition = 0 | 1;

export const Disposition = {
    NUMBER_0: 0 as Disposition,
    NUMBER_1: 1 as Disposition
};